#14 - Energy can now be collected + fixes
- Slots now deselect once a unit has been placed.
- Energy can now be collected

Todo:
- Fix nullreference/missingreference exceptions from enemy coroutines
- Make it so that energy can fall from the sky vertically downward (randomly at more or less fixed intervals)

+++ Assets/Scripts/Entities/Sentinels/Behaviors/Shooter.cs
@@ -44,7 +44,7 @@ public class Shooter : EntityBehavior
         if (!isInitialized)
         {
             var gameData = GameData.Instance;
-            gameData.ChangeEnergy(gameData.GetEnergy() - (int) EnergyCost);
+            gameData.ChangeEnergy(gameData.GetEnergy() - EnergyCost);
             
             yield return StartCoroutine(Initialize());
         }
